I'm going to close this relatively quickly again to get the website updated. This repository isn't too busy yet, so edits can be made in another pull request. It might also be easier to read the developer guidelines when they show up here: http://cms-comp-ops-tools.readthedocs.io/en/latest/developer.html

+++ test/test_build_docs.sh
@@ -4,21 +4,26 @@
 here=`pwd`
 
 # Get the test dir
-testdir=${0%%`basename $0`}
-cd $testdir
+cd ${0%%`basename $0`}
 testdir=`pwd`
 
-# Get directory or package
+if [ "$OPSFULLTEST" != "true" ]
+then
+
+    # Create a virtualenv with your python2.7
+    # Full test already creates one for you
+    virtualenv -p python2.7 venv
+
+    # Put virtualenv in your path
+    PATH=$testdir/venv/bin:$PATH
+
+    # Get rid of PYTHONPATH
+    PYTHONPATH=""
+
+fi
+
+# Get directory of package
 opsdir=${testdir%%"/test"}
-
-# Create a virtualenv with your python2.7
-virtualenv -p python2.7 venv
-
-# Put virtualenv in your path
-PATH=$testdir/venv/bin:$PATH
-
-# Get rid of PYTHONPATH
-PYTHONPATH=""
 
 # Install documentation requirements
 cd $opsdir
@@ -34,7 +39,12 @@ pip install -r docs/requirements.txt
 sphinx-build -b html -E docs test/html
 cd $testdir
 
-# Remove the virtualenv
-rm -rf venv
+if [ "$OPSFULLTEST" != "true" ]
+then
+
+    # Remove the virtualenv
+    rm -rf venv
+
+fi
 
 cd $here
